requirements.txt
Flask==2.1.1
flask-mysqldb==1.0.1
Flask-WTF==0.15.1
Flask-Login==0.5.0
Werkzeug==2.1.1
